update benchmark files

+++ tenspiler/benchmark/cpp/linear_burn_8.cc
@@ -7,14 +7,14 @@
 using namespace std;
 using namespace std::chrono;
 
-vector<vector<float>> linear_burn_8(vector<vector<float>> base, vector<vector<float>> active) {
-    vector<vector<float>> out;
+vector<vector<uint8_t>> linear_burn_8(vector<vector<uint8_t>> base, vector<vector<uint8_t>> active) {
+    vector<vector<uint8_t>> out;
     int m = base.size();
     int n = base[0].size();
     for (int row = 0; row < m; row++) {
-        vector<float> row_vec;
+        vector<uint8_t> row_vec;
         for (int col = 0; col < n; col++) {
-            float pixel = (base[row][col] + active[row][col]) - 255;
+            uint8_t pixel = (base[row][col] + active[row][col]) - 255;
             row_vec.push_back(pixel);
         }
         out.push_back(row_vec);
